Stick to Redhat enterprise version of PHP-5.4
This change is a must. Current legacy code isn't working with currently supported PHP versions (as per https://www.php.net/supported-versions.php)
The PHP-5.x isn't actively maintained. Only RedHat in the RHEL7 actively supports it. We should be using it. So that is why this change is.

Following https://access.redhat.com/solutions/3321431
```
How does Red Hat backport for a version that upstream has already stopped releasing security patches for and where does Red Hat get the fixes from ?
Let's take an example. Say PHP 5.4 is used which has already reached its EOL, so the issue won't be fixed in PHP 5.4 of community version. But for sure it will be fixed in higher versions. Depending upon the severity of the issue, it will be fixed.
```